Fix all clippy warnings (18 → 0)
Applied automatic and manual fixes for all 18 clippy warnings:

**Auto-fixed (12 warnings via cargo clippy --fix):**
- Simplified map_or expressions
- Removed unnecessary type casts
- Replaced or_insert_with with or_insert for simple values
- Used to_vec() instead of iter().copied().collect()
- Collapsed nested if statements
- Fixed !RangeInclusive::contains implementations

**Manually fixed (6 warnings):**
1. Empty doc comment → Regular comment (python.rs:186)
2. Unnecessary unwrap → if let pattern (language_analyzer.rs:201)
3. Manual prefix strip → strip_prefix (language_analyzer.rs:297)
4. Wildcard pattern redundancy → Separate match arms (tdg_tools.rs:95)
5. Manual prefix strip → strip_prefix (hybrid_search.rs:389)
6. Manual prefix strip → strip_prefix (search_engine.rs:341)

**Impact:**
- ✅ Zero clippy warnings remaining
- ✅ All tests passing (35 tests verified)
- ✅ Clean compilation
- ✅ Code quality improved (more idiomatic Rust)
- ✅ Net -2 lines (29 insertions, 31 deletions)

Files modified: 11
- server/src/ast/languages/python.rs
- server/src/cli/language_analyzer.rs
- server/src/mcp/tools/semantic_search_tools.rs
- server/src/mcp_integration/tdg_tools.rs
- server/src/services/enhanced_python_visitor.rs
- server/src/services/hallucination_detector.rs
- server/src/services/semantic/{clustering,hybrid_search,search_engine,topic_modeling}.rs
- server/src/tdg/analyzer_ast.rs

+++ server/src/tdg/analyzer_ast.rs
@@ -1668,7 +1668,7 @@ impl<'a> PythonComplexityVisitor<'a> {
                 // Count parameters
                 if let Some(params) = node.child_by_field_name("parameters") {
                     let param_count = params.child_count();
-                    if param_count > self.max_params as usize {
+                    if param_count > self.max_params {
                         self.max_params = param_count;
                     }
                 }
